Move the HelpLink component one level up.
