+++ js/src/components/contentAnalysis/ReadabilityAnalysis.js
@@ -14,7 +14,7 @@ import Collapsible from "../SidebarCollapsible";
 import getIndicatorForScore from "../../analysis/getIndicatorForScore";
 import { getIconForScore } from "./mapResults";
 import { LocationConsumer } from "../contexts/location";
-import HelpLink from "./HelpLink";
+import HelpLink from "../HelpLink";
 import ReadabilityResultsPortal from "../portals/ReadabilityResultsPortal";
 
 const AnalysisHeader = styled.span`
